requirements.txt
mysql-connector-python==8.1.0
mysqlclient==2.2.0
Pillow==9.5.0